Security 적용, 음식 종류에 상관없이 쿠폰과 기프트카드를 적용할 수 있도록 전체 카테고리 추가

+++ GS24/src/main/java/com/gs24/website/persistence/ImgMapper.java
@@ -4,7 +4,7 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface ImgMapper {
-	long selectNextReviewId();
+	int selectNextReviewId();
 	
-	long selectNextFoodId();
+	int selectNextFoodId();
 }
